network selection changed; task shimmer; task state frame;
