+++ lib/widgets/paw_indicator_with_tasks_list.dart
@@ -9,6 +9,7 @@ import '../blockchain/interface.dart';
 import '../blockchain/task_services.dart';
 import '../task_dialog/task_transition_effect.dart';
 import '../task_item/task_item.dart';
+import '../task_item/task_shimmer.dart';
 
 class PawRefreshAndTasksList extends StatefulWidget {
   final String pageName;
